Upgrading to Teem 1.9. Main changes include API changes for var-arg functions and more careful use of unsigned types.  See http://sourceforge.net/project/shownotes.php?release_id=380683&group_id=85445 for more information.

+++ src/Core/Geom/ColorMappedNrrdTextureObj.cc
@@ -88,8 +88,8 @@ ColorMappedNrrdTextureObj::ColorMappedNrrdTextureObj(NrrdDataHandle &nin,
   nrrd_ = new NrrdData;
 
   if (min_slice != max_slice) {
-    int *min = new int[nin->nrrd->dim];
-    int *max = new int[nin->nrrd->dim];
+    size_t *min = new size_t[nin->nrrd->dim];
+    size_t *max = new size_t[nin->nrrd->dim];
     for (int i = 0; i < nin->nrrd->dim; i++) {
       min[i] = 0;
       max[i] = nin->nrrd->axis[i].size-1;
